requirements.txt
aiogram==2.21
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
async-timeout==4.0.2
attrs==22.1.0
Babel==2.9.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
Deprecated==1.2.13
frozenlist==1.3.1
greenlet==1.1.2
idna==3.3
kombu==5.2.4
multidict==6.0.2
packaging==21.3
prompt-toolkit==3.0.30
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
socialscan==1.4.2
SQLAlchemy==1.4.39
tqdm==4.64.0
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
yarl==1.8.1
